size were written with quint64 and read with int, didnt match up

+++ core/ilwisobjects/domain/colorrange.cpp
@@ -714,7 +714,7 @@ bool ColorPalette::alignWithParent(const IDomain &dom){
 
 void ColorPalette::store(QDataStream &stream)
 {
-    quint64 size = _colors.size();
+    int size = _colors.size();
     stream << size;
     for(const auto& clrItem : _colors)    {
         storeColor(clrItem->color(), stream);
